fix: align entity column constraints with DB schema and externalize API base URL
Add missing precision/scale and length constraints to JPA entities to
match the Flyway migrations, rename User.isActive column to is_ativo,
and read the frontend API base URL from NEXT_PUBLIC_API_URL instead of
hardcoding localhost.

+++ backend/clothing_store/src/main/resources/db/migrations/V12__create_product_images_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE imagens_produtos (
   id_variacao INTEGER       NULL,
   url         TEXT          NOT NULL,
   tipo        VARCHAR(20)   NOT NULL CHECK (tipo IN ('MAIN', 'CAROUSEL')),
-  posicao     SMALLINT      NOT NULL DEFAULT 0,
+  posicao     INTEGER       NOT NULL DEFAULT 0,
   criado_em   TIMESTAMP     NOT NULL DEFAULT now(),
 
   PRIMARY KEY (id),
